Int64 to store value + double rounded

+++ SLPWallet/Extensions/String+Extensions.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 extension String {
-    func toSatoshis() -> Int {
+    func toSatoshis() -> Int64 {
         return self.toDouble().toSatoshis()
     }
